+++ src/Freya.Router/Compilation.fs
@@ -15,6 +15,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+//
 //----------------------------------------------------------------------------
 
 [<AutoOpen>]
@@ -22,74 +23,144 @@ module internal Freya.Router.Compilation
 
 open Aether
 open Aether.Operators
-open Freya.Pipeline
+open FParsec
+open Freya.Core
+open Freya.Types.Uri.Template
+open Hekate
 
-(* Trie *)
+(* Types
 
-type internal CompilationTrie =
-    { Children: CompilationTrie list
-      Key: string
-      Pipelines: (FreyaRouteMethod * FreyaPipeline) list
-      Recognizer: Recognizer }
+   Types representing the elements of a compiled Freya routing graph,
+   modelling each aspect of the graph as a restrictive sum type. *)
 
-    static member ChildrenLens =
-        (fun x -> x.Children), 
-        (fun c x -> { x with Children = c })
+type CompilationGraph =
+    | Graph of Graph<CompilationKey, CompilationNode, CompilationEdge>
 
-    static member PipelinesLens =
-        (fun x -> x.Pipelines), 
-        (fun p x -> { x with Pipelines = p })
+    static member GraphIso : Iso<CompilationGraph, Graph<CompilationKey, CompilationNode, CompilationEdge>> =
+        (fun (Graph g) -> g), (fun g -> Graph g)
 
-and internal Recognizer =
-    | Ignore of string
-    | Capture of string
+and CompilationKey =
+    | Root
+    | Key of string
 
-(* Constructors *)
+and CompilationNode =
+    | Empty
+    | Endpoints of CompilationEndpoint list
 
-let private recognizer (key: string) =
-    match key.StartsWith ":" with
-    | true -> Capture (key.Substring (1))
-    | _ -> Ignore (key)
+and CompilationEndpoint =
+    | Endpoint of FreyaRouteMethod * FreyaPipeline
 
-let private node key =
-    { Children = List.empty
-      Key = key
-      Pipelines = List.empty
-      Recognizer = recognizer key }
+and CompilationEdge =
+    | Edge of Parser<UriTemplateData, unit> * int
 
-let segmentize (path: string) =
-    path.Split '/'
-    |> List.ofArray
-    |> List.filter ((<>) "")
+(* Defaults
 
-(* Lenses *)
+   Default values for common structures, in this case a default (empty)
+   compilation graph for use as the basis in compilation. *)
 
-let private childPLens i =
-    CompilationTrie.ChildrenLens >-?> listPLens i
+let private defaultCompilationGraph =
+    Graph (Graph.create [ Root, Empty ] [])
 
-(* Constructors *)
+(* Lenses
 
-let rec private add node =
-    function | (segment :: path, pipeline, meth) -> 
-                (find segment node |> update segment path pipeline meth) node
-             | (_, pipeline, meth) -> 
-                Lens.map CompilationTrie.PipelinesLens (fun ps -> ps @ [ (meth, pipeline) ]) node
+   Lenses used within compilation to provide access in to the complex
+   data structure(s) used as a routing graph. *)
 
-and private find segment =
-    function | { Children = x } -> List.tryFindIndex (fun x -> x.Key = segment) x
+let graphLens =
+         idLens
+    <--> CompilationGraph.GraphIso
 
-and private update segment path pipeline meth =
-    function | Some i -> extend i path pipeline meth
-             | _ -> append segment path pipeline meth
+(* Patterns
 
-and private extend i path pipeline meth =
-    Lens.mapPartial (childPLens i) (flip add (path, pipeline, meth))
+   Active patterns used to discriminate while compiling a route,
+   distinguishing between a part of the underlying URI Template
+   which forms an intermediate node within the complete route,
+   and the final element which should be represented within
+   the graph as an endpoint (a node which has a non-empty list
+   of Endpoint types). *)
 
-and private append segment path pipeline meth =
-    Lens.map CompilationTrie.ChildrenLens (flip (@) [ add (node segment) (path, pipeline, meth) ])
+let private (|Next|_|) =
+    function | { Method = meth
+                 Specification = spec
+                 Template = UriTemplate (part :: parts)
+                 Pipeline = pipe } -> Some (part, { Method = meth
+                                                    Specification = spec
+                                                    Template = UriTemplate (parts)
+                                                    Pipeline = pipe })
+             | _ -> None
 
-let private addRoute route =
-    (flip add) (segmentize route.Path, route.Pipeline, route.Method)
+let private (|Last|_|) =
+    function | { Method = meth
+                 Specification = spec
+                 Template = UriTemplate (part :: [])
+                 Pipeline = pipe } -> Some (meth, spec, part, pipe)
+             | _ -> None
 
-let compileRoutes =
-    List.fold (flip addRoute) (node "")
+(* Modification
+
+   Functions to modify aspects of the routing graph, chiefly
+   to add routes to the graph (instances of FreyaRoute).
+
+   A fairly simple recurse over the route, taking the head of
+   the URI Template giving the route each time until exhausted. *)
+
+let private composeKeys k1 k2 =
+    match k1, k2 with
+    | Key s1, Key s2 -> Key (s1 + s2)
+    | _, Key s2 -> Key s2
+    | Key s1, _ -> Key s1
+    | _ -> Root
+
+let private addNode key =
+    Graph.addNode (key, Empty)
+
+let private updateNode key meth pipe =
+    Graph.mapNodes (fun key' node ->
+        match key = key' with
+        | true ->
+            match node with
+            | Empty -> Endpoints [ Endpoint (meth, pipe) ]
+            | Endpoints (endpoints) -> Endpoints (endpoints @ [ Endpoint (meth, pipe) ])
+        | _ ->
+            node)
+
+let private addEdge key1 key2 part graph =
+    Graph.addEdge (key1, key2,
+        Edge (UriTemplatePart.Matching.Match part,
+              Option.get (Graph.outwardDegree key1 graph))) graph
+
+let rec private addRoute current graph route =
+    match route with
+    | Last (meth, _, part, pipe) ->
+        let last =
+            composeKeys current (Key (part.ToString ()))
+
+        let graph =
+            ((fun graph ->
+                (match Graph.containsNode last graph with
+                 | false -> addNode last >> updateNode last meth pipe >> addEdge current last part
+                 | _ -> updateNode last meth pipe) graph) ^%= graphLens) graph
+
+        graph
+    | Next (part, route) ->
+        let next =
+            composeKeys current (Key (part.ToString ()))
+
+        let graph =
+            ((fun graph ->
+                (match Graph.containsNode next graph with
+                 | false -> addNode next >> addEdge current next part
+                 | _ -> id) graph) ^%= graphLens) graph
+
+        addRoute next graph route
+    | _ ->
+        graph
+
+(* Compilation
+
+   A function to compile a list of raw FreyaRoute instances to
+   an instance of a CompilationGraph, which can be executed
+   directly (and hopefully efficiently). *)
+
+let compile =
+    List.fold (addRoute Root) defaultCompilationGraph