+++ src/Freya.Types/Parsing.fs
@@ -42,6 +42,12 @@ let ampersandP : Parser<unit, unit> =
 let commaP : Parser<unit, unit> =
     skipChar ','
 
+let dotP : Parser<unit, unit> =
+    skipChar '.'
+
+let equalsP : Parser<unit, unit> =
+    skipChar '='
+
 let semicolonP : Parser<unit, unit> =
     skipChar ';'