+++ RELEASE_NOTES.md
@@ -1,3 +1,13 @@
+### New in 0.11.0-alpha - (Released TBD)
+* Breaking changes:
+  * `Freya.Pipeline` merged into `Freya.Core`
+  * `Freya.Integration` merged into `Freya.Core`
+  * `Freya.Pipeline` and `Freya.Integration` are no longer separate packages
+  * `Freya.Pipeline` is now `Freya.Core.Pipeline` and automatically opened when `Freya.Core` is opened
+  * `Freya.Integration` is its own module named `Freya.Core.Integration`, automatically opened when `Freya.Core` is opened, and no longer split into `Freya.Core.Integration` and `Freya.Pipeline.Integration`
+  * `Freya.Types.Uri.Template` introduced
+  * `Freya.Router` now uses URI Templates as route specifications, along with a new internal graph model
+
 ### New in 0.10.0-alpha - (Released 2015-03-21)
 * Updated Freya.Types.* family
 * Naming convention updates and simplification